Safari-611.1.10/LayoutTests/accessibility/mac/list-items-ignored-expected.txt


Mar 7
test
test


Mar 7
test
test


Mar 7
test
test

This tests that list items will not be ignored as children of a list.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS list.childrenCount is 3
PASS successfullyParsed is true

TEST COMPLETE
